fix: should not relay on buissness logic in block std
