+++ packages/blocks/src/page-block/page-service.ts
@@ -1,4 +1,5 @@
 import { BlockService } from '@blocksuite/block-std';
+import { assertExists } from '@blocksuite/global/utils';
 import type { BlockElement } from '@blocksuite/lit';
 
 import {
@@ -23,8 +24,10 @@ import {
   getTextSelectionCommand,
   withHostCommand,
 } from './commands/index.js';
+import type { EdgelessPageBlockComponent } from './edgeless/edgeless-page-block.js';
 import { FontLoader } from './font-loader/font-loader.js';
 import type { PageBlockModel } from './page-model.js';
+import type { PageBlockComponent } from './types.js';
 
 export class PageService extends BlockService<PageBlockModel> {
   readonly fontLoader = new FontLoader();
@@ -35,8 +38,65 @@ export class PageService extends BlockService<PageBlockModel> {
     flavour: this.flavour,
   };
 
+  get viewportElement() {
+    const pageElement = this.std.view.viewFromPath('block', [
+      this.std.page.root?.id ?? '',
+    ]) as PageBlockComponent | EdgelessPageBlockComponent | null;
+    assertExists(pageElement);
+    const viewportElement = pageElement.viewportElement as HTMLElement | null;
+    assertExists(viewportElement);
+    return viewportElement;
+  }
+
   override mounted() {
     super.mounted();
+
+    this.handleEvent(
+      'selectionChange',
+      () => {
+        const viewport = this.viewportElement;
+        if (!viewport) return;
+
+        const selection = document.getSelection();
+        if (!selection || !selection.rangeCount) return;
+        const range = selection.getRangeAt(0);
+
+        const startElement =
+          range.startContainer instanceof Element
+            ? range.startContainer
+            : range.startContainer.parentElement;
+        if (!startElement) return;
+
+        if (
+          startElement === document.documentElement ||
+          startElement === document.body
+        )
+          return;
+
+        if (startElement.closest('.blocksuite-overlay')) return;
+
+        if (!viewport.contains(startElement)) {
+          this.std.event.deactivate();
+          return;
+        }
+
+        const endElement =
+          range.endContainer instanceof Element
+            ? range.endContainer
+            : range.endContainer.parentElement;
+        if (!endElement) return;
+
+        if (!viewport.contains(endElement)) {
+          this.std.event.deactivate();
+          return;
+        }
+
+        this.std.event.activate();
+      },
+      {
+        global: true,
+      }
+    );
 
     this.std.command
       .add('getBlockIndex', getBlockIndexCommand)
